requirements.txt
asgiref==3.6.0
autopep8==2.0.2
backports.zoneinfo==0.2.1
Django==4.2.1
django-cors-headers==3.14.0
django-dotenv==1.4.2
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Faker==18.7.0
phonenumbers==8.13.11
phonenumberslite==8.13.11
Pillow==9.5.0
pycodestyle==2.10.0
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
